Machine_Learning_Engineering/Deployment_to_Azure_Kubernetes/pyproject.toml
[tool.poetry]
name = "lab3"
version = "0.3.0"
description = ""
authors = ["Anson Quon <aquon@berkeley.edu>"]

[tool.poetry.dependencies]
python = "~3.11"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
joblib = "^1.3.2"
scikit-learn = "1.3.2"
pydantic = "~2.5"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.9"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
httpx = "^0.26"
ruff = "^0.1"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
